Convert hkdf::Algorithm::hmac_algorithm and hmac::Algorithm::digest_algorithm to const fns
Currently, its impossible to have a const fn in a trait, so the len()
method of an HMAC algorithm from its KeyType trait implementation
can't be used in consts and statics.

This works around that by making it so you can chain together
hmac_algorithm().digest_algorithm() to obtain the output_len in a constant.